requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
bcrypt==4.1.3
pydicom==2.4.4
numpy==1.26.4
cryptography==42.0.8
Werkzeug==3.0.3
python-dotenv==1.0.0
click==8.1.7
flask-jwt-extended==4.6.0
pydantic[email]==2.6.4
flasgger==0.9.7.1
matplotlib==3.8.4
scipy==1.13.0
